moved test-contingency
